Refactored Okta API to organize Okta API functions more logically.
